Stronger typing for slots
Use a union instead of casts. We now use casts very sparingly.

The initial goal was to make slots volatile. This would allow us to
distinguish real roots from fake roots by looking at the shape of the
predecessor. For fake roots we would place a sentinel value. This
approach does not work due to compaction in OCaml 4:

- We cannot force slot access to be volatile: the root needs to
  survive compaction so we cannot use a proxy, (in OCaml 5 on the
  other hand it is already volatile)

- We cannot make assumptions on the shape of values anyway: compaction
  replaces values with arbitrary contents during pointer inversion.

Also, it is not excluded that compaction comes back in OCaml 5 at some
point.