Bitmap allocator
bitmap_boxroot is inspired by the Hotspot VM implementation of JNI
global references using a bitmap allocator of 64-sized chunks. For
benchmarking purposes.

Allocation uses a lock and modifies the bitmap atomically whereas
deallocation is lock-free most of the time. We use a simple
generational strategy, without which we already know the performance
is bad, since I do not know the Hotspot GC enough for an
apples-to-apples comparison on this aspect.