Nth micro-optim to boxroot: smarter demotion
When a pool is empty after a deletion, we move it to the free pools,
but not if the pool is current, to avoid repeated pathological cases.
This PR uses a trick to avoid taking the slow path in the pathological
cases.

The pathological case is observed with the local_roots benchmark,
which records significant gains for low N.